refactor: move test data to mocks
