+++ package.json
@@ -102,7 +102,8 @@
     "exclude": [
       "cypress/**/*",
       "!cypress/integration/e2e/**/*",
-      "__mocks__/**/*"
+      "__mocks__/**/*",
+      "**/*.test.tsx"
     ]
   }
 }
